Seperated keyboard-related code from the ManusVibrate class.
