+++ Assets/Scripts/KeyBoardController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 0067ca6c38c569a46831c327241af6c2
+timeCreated: 1465306289
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
